refactor(codegen): generate code with Java 8 and improve the namespacing
